+++ ChangeLog.md
@@ -2,7 +2,7 @@
 
 ## [Unrelease]
 ### Added
-- Add the python code `generatePkgConfig.py` to use the arguments from command line to find the include path and library files for libtorch.
+- Add the python code `generatePkgConfig.py` to generate `.pc` file for the command `pkg-config`.
 
 ### Fixed